refactor(pkg/trace/stats): move pkg/trace/stats/otel.go to pkg/trace/otel/stats
